fix: address flashcard plan pr review feedback
