+++ apps/packages/ui/src/components/Flashcards/hooks/useFlashcardQueries.ts
@@ -1014,17 +1014,19 @@ export function useGenerateFlashcardsMutation() {
       focusTopics?: string[]
       provider?: string
       model?: string
-    }) =>
-      generateFlashcards({
+    }) => {
+      const hasCardPlan = Array.isArray(params.cardPlan) && params.cardPlan.length > 0
+      return generateFlashcards({
         text: params.text,
         num_cards: params.numCards,
-        card_type: params.cardPlan ? undefined : params.cardType,
-        card_plan: params.cardPlan,
+        card_type: hasCardPlan ? undefined : params.cardType,
+        card_plan: hasCardPlan ? params.cardPlan : undefined,
         difficulty: params.difficulty,
         focus_topics: params.focusTopics,
         provider: params.provider,
         model: params.model
-      }),
+      })
+    },
     onError: (error) => {
       console.error("Failed to generate flashcards:", error)
     }
